DISTRO: add OPENSTLINUX_RELEASE flag to DISTRO_VERSION
Set by default OPENSTLINUX_RELEASE to 'snapshot', in this case:
  DISTRO_VERSION = "2.2+snapshot-${DATE}"
  SDK_VERSION = "2.2+snapshot"
If OPENSTLINUX_RELEASE is set to official tag, e.g. openstlinux-17-07-07
  DISTRO_VERSION = "2.2+openstlinux-17-07-07"
  SDK_VERSION = "2.2+openstlinux-17-07-07"
NB: this is the way of working with OPENSTLINUX_RELEASE, as soon as it is
    defined and it is not having 'snapshot' string.

By updating DISTRO_VERSION with OPENSTLINUX_RELEASE flag, we can get this
info displayed on target throught 'lsb_release'.
And it is also displayed at build time on console throught 'BUILDCFG_VARS'

Change-Id: Id3b5a4b4cdbe4960493bfba8d6da8c23d6b3245f
Signed-off-by: Romuald JEANNE <romuald.jeanne@st.com>
Reviewed-on: https://gerrit.st.com/75551
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Sebastien GANDON <sebastien.gandon@st.com>