GSTREAMER: disable waylandsink redraw callback
This workaround is needed to reach 30fps video playback
up to display side.

Redraw callback is received late, leading to many frames
being dropped and so low framerate on display driver stage.
The fact that redraw callback is received late is not yet understood
but display subsystem is able to sustain a higher framerate
than the one reached with redraw callback mechanism.
This workaround consists to ignore redraw callback allowing to
get back to the expected framerate performances.

Even if not observed yet, there is a risk of visible decoding artefacts
due to the fact that downstream elements could rewrite in a frame
being currently rendered by display subsystem...

Change-Id: I6b56d10c6c90958265a2817dce813dd0e437d5a9
Signed-off-by: Hugues Fruchet <hugues.fruchet@st.com>
Reviewed-on: https://gerrit.st.com/114096
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Benjamin GAIGNARD <benjamin.gaignard@st.com>
Reviewed-by: Philippe CORNU <philippe.cornu@st.com>
Reviewed-by: Christophe MONTAUD <christophe.montaud@st.com>