DEMO: add AI Character Recognition Launcher
This recipe add the support of the AI Character Recognition Launcher
application.

This example demonstrate a complex application that is running on
both CPU1(CA7) and CPU2(CM4).
The application is a launcher that recognize hand writing character
drawn on the touch screen in order to execute specific actions.

How to launch it:
/usr/local/demo/bin/ai_char_reco_launcher /usr/local/demo/bin/apps_launcher_example.sh

Change-Id: I7f2de330e1ee56d71d2b02b458187803beab323a
Signed-off-by: Vincent Abriou <vincent.abriou@st.com>
Reviewed-on: https://gerrit.st.com/114690
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Christophe MONTAUD <christophe.montaud@st.com>