DEMO-LAUNCHER: fix wrong declared dependency on dynamic libraries
Currently 'weston-simple-st-egl-tex' file set 'libEGL.so' and
'libGLESv2.so' in NEEDED section.
This works fine with specific EGL and LIBGLES2 providers such as
'gcnano-userland'.
But this is not the correct setting for generic use case, when, for example,
EGL and LIBGLES2 are provided by 'mesa'

Fix this issue by updating 'weston-simple-st-egl-tex' binary with expected
library name version when 'mesa' is the preferred provider.

Change-Id: I7675e17af2ffc673858c5cd384383d4a0bf83891
Signed-off-by: Romuald JEANNE <romuald.jeanne@st.com>
Reviewed-on: https://gerrit.st.com/116206
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Christophe MONTAUD <christophe.montaud@st.com>