CLASSES: fix issue in reproducible_build with 'devtool add' command
Specific use case for 'devtool add' recipe
Temporary fix as the source-date-epoch folder is not yet handled properly
in 'create.py' file:
  (line 507)      filterout = ['git.indirectionsymlink']
This has to be fixed with:
                  filterout = ['git.indirectionsymlink', 'source-date-epoch']
But 'source-date-epoch' may have other name, so need to make sure to use
the right one.
We propose a workaround in this class to avoid patching 'create.py' file.
At the end, the 'do_create_source_date_epoch_stamp' should not be called
when expecting to add a new recipe with 'devtool add'.

Change-Id: Ia2740f0918f20473cbf0a8495478dab972c28af0
Signed-off-by: Romuald JEANNE <romuald.jeanne@st.com>
Reviewed-on: https://gerrit.st.com/116661
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Christophe MONTAUD <christophe.montaud@st.com>