DEMO: ai-hand-char-reco-launcher: keep all action on the same layer
When launching different action, like video and camera in parallel,
both are visible within the same layer. It avoid the grey transparent
effect that we had before.
Further, the loop to check the touch screen input device has been
improve to be more robust in case a event device is remove from the
list.

Change-Id: I82e86bb3e5c766a2c412e46c672e930863367339
Signed-off-by: Vincent Abriou <vincent.abriou@st.com>
Reviewed-on: https://gerrit.st.com/116814
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Christophe MONTAUD <christophe.montaud@st.com>