DEMO: ai-hand-char-reco-launcher: Fix to avoid M4 firmware RPMsg corruption
The M4 firmware has an issue of buffer overflow. The 28x28 buffer
have sometime an overflow of one pixel due to the drawing pitch of 2
implemented in the M4 firmware to ensure proper character detection by
the Neural Network.
To avoid this the draw area size sent by the host to the corpo has been
reduced from 28x28 to 25x25. This trick avoid the buffer overflow on the
M4 size.
The real fix need however to be implemented in the M4 firmware.

Signed-off-by: Vincent Abriou <vincent.abriou@st.com>
Change-Id: I154aa6bf1abc20a4a2dc1d76691bab99fa29e542
Reviewed-on: https://gerrit.st.com/125009
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Christophe PRIOUZEAU <christophe.priouzeau@st.com>