ltrace: Add support for ARM Thumb branches with instruction-set exchange.
This is issue has been detected on OpenSTLinux 2019-10-09, where
ltrace failed to single-step on a "bx lr" instruction when tracing
weston-simple-egl.

The original maintainer of ltrace has been contacted without success,
and the projet hosted on Github is not active anymore.  As a
consequence, this patch should be submitted to the Yocto project.

Fixes: https://intbugzilla.st.com/show_bug.cgi?id=69540
Change-Id: Iedb033ac02f9b622cac1f6abc93e56750a4e1e63
Reviewed-on: https://gerrit.st.com/c/oeivi/oe/st/meta-opensdk/+/155271
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Christophe PRIOUZEAU <christophe.priouzeau@st.com>
Tested-by: Christophe PRIOUZEAU <christophe.priouzeau@st.com>