QT: EGLFS: add screen 'size' field in json config
While by default eglfs uses the display in fullscreen, it is
possible to force QT to use only one part of the whole display
through the 'size' parameter in the json configuration script.

Add the 'size' parameter equal to the whole display size, just as
placehold for a user's value.

Unfortunately it's not possible to add comments in a json file to
better guide the user.

Signed-off-by: Antonio Borneo <antonio.borneo@st.com>
Change-Id: I7c33c24a763958c32955364e1c936f445ca40e4d
Reviewed-on: https://gerrit.st.com/c/oeivi/oe/st/meta-opensdk/+/178452
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Christophe PRIOUZEAU <christophe.priouzeau@st.com>
Tested-by: Christophe PRIOUZEAU <christophe.priouzeau@st.com>