DEMO-AI : improve display and input detection
a) automatic input detection : if a mouse is detected,
it is used even if a touchscreen is there.
If mouse is not plugged, touchscreen is default input.

b) automatic display resolution : the display output is read
from the connected device from modetest command, and resolution
is read from /sys/class/drm/card0-<display>.

The mouse returns relative coordinates, which needs to be
converted into absolutes, and after a while, an offset
between cursor given by Wayland and the computed coordinates could
happend, especially when cursor wants to go outside the screen.
That's why mouse cursor is initialized at each draw.

An improvement proposition would to use libinput which is the
layer between kernel and Wayland which manages inputs, but the
application will depend on Wayland.
Here is a piece of code as example if we want to go on that direction :
https://github.com/wayland-project/libinput/blob/master/test/test-touch.c

Change-Id: I6a4db4512db5210ae94bf730ff364d97a9317476
Reviewed-on: https://gerrit.st.com/c/oeivi/oe/st/meta-opensdk/+/197207
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Christophe GUIBOUT <christophe.guibout@st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Christophe PRIOUZEAU <christophe.priouzeau@st.com>
Tested-by: Christophe GUIBOUT <christophe.guibout@st.com>